refator: refactored prompt converting files for cohere and jamba and extracted helper functions to separate files for much neater code
